+++ src/plugins/e-acsl/tests/bts/oracle/gen_issue-eacsl-172.c
@@ -12,7 +12,7 @@ double d2 = 11.;
 int main(void)
 {
   int __retres;
-  __e_acsl_memory_init((int *)0,(char ***)0,(size_t)8);
+  __e_acsl_memory_init((int *)0,(char ***)0,8UL);
   {
     __e_acsl_assert_data_t __gen_e_acsl_assert_data = {.values = (void *)0};
     __e_acsl_assert_register_double(& __gen_e_acsl_assert_data,"d2",d2);
