+++ src/plugins/e-acsl/tests/memory/oracle/gen_errno.c
@@ -15,7 +15,7 @@ void __e_acsl_globals_init(void)
   static char __e_acsl_already_run = 0;
   if (! __e_acsl_already_run) {
     __e_acsl_already_run = 1;
-    __e_acsl_store_block((void *)(& errno),(size_t)4);
+    __e_acsl_store_block((void *)(& errno),4UL);
     __e_acsl_full_init((void *)(& errno));
   }
   return;
@@ -30,10 +30,10 @@ void __e_acsl_globals_clean(void)
 int main(int argc, char const **argv)
 {
   int __retres;
-  __e_acsl_memory_init(& argc,(char ***)(& argv),(size_t)8);
+  __e_acsl_memory_init(& argc,(char ***)(& argv),8UL);
   __e_acsl_globals_init();
   int *p = & errno;
-  __e_acsl_store_block((void *)(& p),(size_t)8);
+  __e_acsl_store_block((void *)(& p),8UL);
   __e_acsl_full_init((void *)(& p));
   {
     int __gen_e_acsl_initialized;
